test: move util test to test/lib/UtilTest.py
