+++ .travis.yml
@@ -24,7 +24,7 @@ script:
   - cd test/lib && python -m unittest -v UtilTest
   - python -m unittest -v VisTest
   - cd analysis && python -m unittest -v ChannelTest
-  - cd ../.. && python -m unittest -v unit_tests && python -m unittest -v tests
+  - cd ../.. && python -m unittest -v tests
   - cd slack && python -m unittest -v unit_tests
   - cd lib && python -m unittest -v UtilTest
   - cd analysis && python -m unittest -v ChannelTest
